refactor: move import tests to separate file
